refactor(RepositoryTest): 기존 테스트 코드 통합 테스트 패키지로 이동
