+++ src/test/java/gift/repository/integration/JpaProductRepositoryTest.java
@@ -1,10 +1,11 @@
-package gift.repository;
+package gift.repository.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import gift.domain.Product;
+import gift.repository.JpaProductRepository;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.junit.jupiter.api.BeforeEach;
