+++ src/test/java/gift/repository/integration/JpaUserRepositoryTest.java
@@ -1,10 +1,11 @@
-package gift.repository;
+package gift.repository.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import gift.domain.User;
+import gift.repository.JpaUserRepository;
 import java.util.NoSuchElementException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
